api/requirements.txt
Flask
gunicorn
python-dotenv
earthengine-api
google-api-python-client
